+++ src/avdecc/helper.cpp
@@ -1049,6 +1049,36 @@ QString loggerLevelToString(la::avdecc::logger::Level const& level) noexcept
 	}
 }
 
+QString generateDumpSourceString(QString const& shortName, QString const& version) noexcept
+{
+	static auto s_DumpSource = QString{ "%1 v%2 using L-Acoustics AVDECC Controller v%3" }.arg(shortName).arg(version).arg(la::avdecc::controller::getVersion().c_str());
+
+	return s_DumpSource;
+}
+
+bool isValidEntityModelID(la::avdecc::UniqueIdentifier const entityModelID) noexcept
+{
+	if (entityModelID)
+	{
+		auto const [vendorID, deviceID, modelID] = la::avdecc::entity::model::splitEntityModelID(entityModelID);
+		return vendorID != 0x00000000 && vendorID != 0x00FFFFFF;
+	}
+	return false;
+}
+
+bool isEntityModelComplete(la::avdecc::UniqueIdentifier const entityID) noexcept
+{
+	auto& manager = hive::modelsLibrary::ControllerManager::getInstance();
+	auto controlledEntity = manager.getControlledEntity(entityID);
+
+	if (controlledEntity)
+	{
+		return controlledEntity->isEntityModelValidForCaching();
+	}
+
+	return true;
+}
+
 void smartChangeInputStreamFormat(QWidget* const parent, bool const autoRemoveMappings, la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::StreamIndex const streamIndex, la::avdecc::entity::model::StreamFormat const streamFormat, QObject* context, std::function<void(hive::modelsLibrary::CommandsExecutor::ExecutorResult const result)> const& handler) noexcept
 {
 	AVDECC_ASSERT(context != nullptr, "context must not be nullptr");